Revert "Aml 2372 losing paye scheme references"

+++ src/SFA.DAS.EAS.Employer_Financial.Database/Tables/AccountTransfers.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [employer_financial].[AccountTransfers]
 (
-    [Id] BIGINT NOT NULL PRIMARY KEY IDENTITY, 
+    [Id] INT NOT NULL PRIMARY KEY IDENTITY, 
     [SenderAccountId] BIGINT NOT NULL, 
     [SenderAccountName] NVARCHAR(100) NOT NULL, 
     [ReceiverAccountId] BIGINT NOT NULL,
